Refactor main app into multiple files
